Move documentation to Next Fumadocs app
