+++ .github/workflows/deploy.yml
@@ -23,8 +23,16 @@ jobs:
       - name: Check
         run: bun run check
 
-      - name: Deploy to Cloudflare Workers
+      - name: Build docs for Cloudflare
+        run: bun run build:docs
+
+      - name: Deploy API Worker to Cloudflare
         uses: cloudflare/wrangler-action@v3
         with:
           apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           command: deploy --minify src/index.ts
+
+      - name: Deploy docs Worker to Cloudflare
+        run: bun run deploy:docs
+        env:
+          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
