+++ apps/docs/app/api/search/route.ts
@@ -1,0 +1,1 @@
+export { GET } from "../../../../../src/docs/responses";
