+++ apps/docs/app/globals.css
@@ -1,0 +1,25 @@
+@import "fumadocs-ui/style.css";
+
+.endpoint-summary {
+  margin-bottom: 2rem;
+  border-radius: 0.75rem;
+  border: 1px solid var(--color-fd-border);
+  background: var(--color-fd-card);
+  padding: 1rem;
+}
+
+.endpoint-summary p {
+  margin: 0.5rem 0;
+}
+
+.endpoint-method {
+  display: inline-block;
+  margin-right: 0.5rem;
+  border-radius: 999px;
+  background: var(--color-fd-primary);
+  color: var(--color-fd-primary-foreground);
+  padding: 0.125rem 0.5rem;
+  font-size: 0.75rem;
+  font-weight: 700;
+  letter-spacing: 0.06em;
+}
