+++ apps/docs/next.config.mjs
@@ -1,0 +1,6 @@
+/** @type {import("next").NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+};
+
+export default nextConfig;
