+++ apps/docs/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@rates-api/docs",
+  "private": true,
+  "scripts": {
+    "build": "next build",
+    "build:cloudflare": "opennextjs-cloudflare build",
+    "dev": "next dev",
+    "preview": "opennextjs-cloudflare preview",
+    "deploy": "opennextjs-cloudflare deploy"
+  }
+}
