+++ apps/docs/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "baseUrl": "../..",
+    "ignoreDeprecations": "6.0",
+    "incremental": true,
+    "jsx": "preserve",
+    "noEmit": true,
+    "paths": {
+      "@/*": ["apps/docs/*"],
+      "@rates-api/*": ["src/*"]
+    },
+    "plugins": [{ "name": "next" }]
+  },
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "../../src/docs/**/*.ts"
+  ],
+  "exclude": ["node_modules"]
+}
