+++ apps/docs/wrangler.toml
@@ -1,0 +1,8 @@
+name = "ratesapi-nz-docs"
+main = ".open-next/worker.js"
+compatibility_date = "2025-06-01"
+compatibility_flags = ["nodejs_compat"]
+
+[assets]
+directory = ".open-next/assets"
+binding = "ASSETS"
