+++ tsconfig.json
@@ -30,7 +30,12 @@
     "lib": ["ESNext", "DOM"],
     "types": ["@cloudflare/workers-types", "bun-types"],
     "typeRoots": ["./node_modules"],
+    "paths": {
+      "@/*": ["./apps/docs/*"],
+      "@rates-api/*": ["./src/*"]
+    },
 
     "noErrorTruncation": true // TLDR - Don't truncate errors (useful for debugging)
-  }
+  },
+  "exclude": ["apps/docs/.next", "apps/docs/**/*.ts", "apps/docs/**/*.tsx"]
 }
